refector: chatting sync modify
