+++ src/apis/conversation/dtos.ts
@@ -25,3 +25,26 @@ export interface PreviousRoomResponseDto {
   success: boolean;
   data: Data;
 }
+
+export class ChattingSocketResponse {
+  name: string;
+  message: string;
+  email: string;
+  date: Date;
+  constructor({
+    name,
+    message,
+    email,
+    date,
+  }: {
+    name: string;
+    message: string;
+    email: string;
+    date: string;
+  }) {
+    this.name = name;
+    this.message = message;
+    this.email = email;
+    this.date = new Date(date);
+  }
+}
